Finally bug fixed with inputs
When the user had to input a data in the new hotspot created, for some
reason, only the first input was displayed, and when the user left,
naturally only the title got updated.

The problem was that the function written to show the inputs, was
getting and showing exclusively the first input.