Hotspot creator improvements
Hotspot creator has better search queries for the elements that will be
modified.

It has edge detection, so if the user plots to close to the sides, the
indicator will be moved, to not overflow x.

The card won't overflow x, based on the window size, if it is out of the
safe area it will be moved out left or right.